Reorganize inference4j-core into session, model, and processing sub-packages
Move 9 types out of the flat root package into domain-focused sub-packages
to improve discoverability as the project grows:

- io.github.inference4j.session: SessionConfigurer, SessionOptions
- io.github.inference4j.model: ModelSource, HuggingFaceModelSource, LocalModelSource
- io.github.inference4j.processing: Preprocessor, Postprocessor, OutputOperator, MathOps

Core contracts (InferenceTask, Classifier, Detector, AbstractInferenceTask,
Tensor, TensorType, InferenceSession, InferenceContext) remain at root since
they are the most-imported types across all modules.

SessionOptions.toOrtOptions() promoted to public to allow cross-package
access from InferenceSession.
