+++ inference4j-core/src/main/java/io/github/inference4j/AbstractInferenceTask.java
@@ -16,6 +16,9 @@
 
 package io.github.inference4j;
 
+import io.github.inference4j.processing.Postprocessor;
+import io.github.inference4j.processing.Preprocessor;
+
 import java.util.Map;
 
 /**
