+++ inference4j-core/src/main/java/io/github/inference4j/model/HuggingFaceModelSource.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.github.inference4j;
+package io.github.inference4j.model;
 
 import io.github.inference4j.exception.ModelDownloadException;
 import org.slf4j.Logger;
