+++ inference4j-core/src/main/java/io/github/inference4j/model/ModelSource.java
@@ -14,7 +14,9 @@
  * limitations under the License.
  */
 
-package io.github.inference4j;
+package io.github.inference4j.model;
+
+import io.github.inference4j.InferenceSession;
 
 import java.nio.file.Path;
