+++ inference4j-core/src/main/java/io/github/inference4j/processing/OutputOperator.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.github.inference4j;
+package io.github.inference4j.processing;
 
 import java.util.Objects;
