+++ inference4j-core/src/test/java/io/github/inference4j/processing/OutputOperatorTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.github.inference4j;
+package io.github.inference4j.processing;
 
 import org.junit.jupiter.api.Test;
