+++ inference4j-tasks/src/main/java/io/github/inference4j/nlp/DistilBertTextClassifier.java
@@ -17,12 +17,12 @@
 package io.github.inference4j.nlp;
 
 import io.github.inference4j.AbstractInferenceTask;
-import io.github.inference4j.HuggingFaceModelSource;
+import io.github.inference4j.model.HuggingFaceModelSource;
 import io.github.inference4j.InferenceSession;
-import io.github.inference4j.MathOps;
-import io.github.inference4j.ModelSource;
-import io.github.inference4j.OutputOperator;
-import io.github.inference4j.SessionConfigurer;
+import io.github.inference4j.processing.MathOps;
+import io.github.inference4j.model.ModelSource;
+import io.github.inference4j.processing.OutputOperator;
+import io.github.inference4j.session.SessionConfigurer;
 import io.github.inference4j.Tensor;
 import io.github.inference4j.exception.ModelSourceException;
 import io.github.inference4j.text.ModelConfig;
@@ -135,7 +135,7 @@ public class DistilBertTextClassifier
         return results;
     }
 
-    private static io.github.inference4j.Preprocessor<String, Map<String, Tensor>> createPreprocessor(
+    private static io.github.inference4j.processing.Preprocessor<String, Map<String, Tensor>> createPreprocessor(
             Tokenizer tokenizer, int maxLength, Set<String> expectedInputs) {
         return text -> {
             EncodedInput encoded = tokenizer.encode(text, maxLength);
