+++ inference4j-tasks/src/main/java/io/github/inference4j/nlp/MiniLMSearchReranker.java
@@ -17,11 +17,11 @@
 package io.github.inference4j.nlp;
 
 import io.github.inference4j.AbstractInferenceTask;
-import io.github.inference4j.HuggingFaceModelSource;
+import io.github.inference4j.model.HuggingFaceModelSource;
 import io.github.inference4j.InferenceSession;
-import io.github.inference4j.ModelSource;
-import io.github.inference4j.Preprocessor;
-import io.github.inference4j.SessionConfigurer;
+import io.github.inference4j.model.ModelSource;
+import io.github.inference4j.processing.Preprocessor;
+import io.github.inference4j.session.SessionConfigurer;
 import io.github.inference4j.Tensor;
 import io.github.inference4j.exception.ModelSourceException;
 import io.github.inference4j.tokenizer.EncodedInput;
