+++ inference4j-tasks/src/main/java/io/github/inference4j/nlp/SentenceTransformerEmbedder.java
@@ -17,10 +17,10 @@
 package io.github.inference4j.nlp;
 
 import io.github.inference4j.AbstractInferenceTask;
-import io.github.inference4j.HuggingFaceModelSource;
+import io.github.inference4j.model.HuggingFaceModelSource;
 import io.github.inference4j.InferenceSession;
-import io.github.inference4j.ModelSource;
-import io.github.inference4j.SessionConfigurer;
+import io.github.inference4j.model.ModelSource;
+import io.github.inference4j.session.SessionConfigurer;
 import io.github.inference4j.Tensor;
 import io.github.inference4j.exception.ModelSourceException;
 import io.github.inference4j.tokenizer.EncodedInput;
@@ -127,7 +127,7 @@ public class SentenceTransformerEmbedder
         };
     }
 
-    private static io.github.inference4j.Preprocessor<String, Map<String, Tensor>> createPreprocessor(
+    private static io.github.inference4j.processing.Preprocessor<String, Map<String, Tensor>> createPreprocessor(
             Tokenizer tokenizer, int maxLength, Set<String> expectedInputs) {
         return text -> {
             EncodedInput encoded = tokenizer.encode(text, maxLength);
