+++ inference4j-tasks/src/main/java/io/github/inference4j/vision/EfficientNetClassifier.java
@@ -16,11 +16,11 @@
 
 package io.github.inference4j.vision;
 
-import io.github.inference4j.HuggingFaceModelSource;
+import io.github.inference4j.model.HuggingFaceModelSource;
 import io.github.inference4j.InferenceSession;
-import io.github.inference4j.ModelSource;
-import io.github.inference4j.OutputOperator;
-import io.github.inference4j.Preprocessor;
+import io.github.inference4j.model.ModelSource;
+import io.github.inference4j.processing.OutputOperator;
+import io.github.inference4j.processing.Preprocessor;
 import io.github.inference4j.Tensor;
 import io.github.inference4j.exception.ModelSourceException;
 import io.github.inference4j.image.Labels;
