+++ inference4j-tasks/src/main/java/io/github/inference4j/vision/YoloV8Detector.java
@@ -17,12 +17,12 @@
 package io.github.inference4j.vision;
 
 import io.github.inference4j.AbstractInferenceTask;
-import io.github.inference4j.HuggingFaceModelSource;
+import io.github.inference4j.model.HuggingFaceModelSource;
 import io.github.inference4j.InferenceContext;
 import io.github.inference4j.InferenceSession;
-import io.github.inference4j.MathOps;
-import io.github.inference4j.ModelSource;
-import io.github.inference4j.SessionConfigurer;
+import io.github.inference4j.processing.MathOps;
+import io.github.inference4j.model.ModelSource;
+import io.github.inference4j.session.SessionConfigurer;
 import io.github.inference4j.Tensor;
 import io.github.inference4j.exception.InferenceException;
 import io.github.inference4j.exception.ModelSourceException;
@@ -161,7 +161,7 @@ public class YoloV8Detector
 
     // --- Preprocessing ---
 
-    private static io.github.inference4j.Preprocessor<BufferedImage, Map<String, Tensor>> createPreprocessor(
+    private static io.github.inference4j.processing.Preprocessor<BufferedImage, Map<String, Tensor>> createPreprocessor(
             String inputName, int inputSize) {
         return image -> {
             LetterboxResult lb = letterbox(image, inputSize);
