+++ inference4j-tasks/src/test/java/io/github/inference4j/nlp/DistilBertTextClassifierTest.java
@@ -17,8 +17,8 @@
 package io.github.inference4j.nlp;
 
 import io.github.inference4j.InferenceSession;
-import io.github.inference4j.ModelSource;
-import io.github.inference4j.OutputOperator;
+import io.github.inference4j.model.ModelSource;
+import io.github.inference4j.processing.OutputOperator;
 import io.github.inference4j.Tensor;
 import io.github.inference4j.exception.ModelSourceException;
 import io.github.inference4j.tokenizer.EncodedInput;
